requirements.txt
google-cloud-storage
google-cloud-aiplatform
cloud-sql-python-connector
vertexai
python-dotenv
asyncio
typing
google-auth
SQLAlchemy
aiohttp
asyncpg
greenlet
pg8000